chore(feeds): Code cleanup
